Get everything (package wrapper)

Get everything (package wrapper)

Usage

getEverything(occs, vars = "all", years = 1979:2013, region = NULL, res = 5)

Arguments

occs

species occurrence coordinates (2 columns in this order: x, y or LONgitude, LATitude) in an object coercible to a data.frame (e.g. a data.frame, matrix, tibble, sf object or SpatVector of points), and in the same coordinate reference system as 'rasts'

vars

character vector of the names of the variables to use - see varsAvailable()

years

year range for the model time series; default 1979:2013

region

optional SpatExtent or SpatVector polygon delimiting the region of 'rasts' within which to compute the models; see ?getRegion for suggestions

res

spatial resolution (pixel size), if larger than the original variable raster layers', in which case terra::aggregate() is used

Author

A. Marcia Barbosa