Get region

This function suggests a region for building an ecological niche model around a given set of species occurrence point coordinates. Mind that this region does not consider survey effort, geographical barriers or other factors that should also be taken into account when delimiting a region for modelling.

Usage

getRegion(occs, type = "mean", prop = 0.5, crs = "EPSG:4326")

Arguments

occs

species occurrence coordinates (2 columns in this order: x, y or LONgitude, LATitude) in an object coercible to a data.frame (e.g. a data.frame, matrix, tibble, sf object or SpatVector of points)

type

character indicating whether to use the "mean" pairwise terra::distance() among points (the default), or a proportion of the "width" (minimum diameter, computed with terra::width()) of the points' spatial extent

prop

if type="width", proportion of the width to use for the terra::buffer() radius. Default 0.5

crs

coordinate reference system behind the coordinates of 'occs', in one of the following formats: WKT/WKT2, authority:code (e.g. EPSG:4326), or PROJ-string notation (see terra::crs()). The default is EPSG:4326, i.e. unprojected longitude-latitude coordinates in geodetic degrees, datum WGS 1984.

See also

Author

A. Marcia Barbosa